+++ src/pfjax/__init__.py
@@ -1,3 +1,3 @@
 from .simulate import *
-from .full_loglik import *
+from .loglik_full import *
 from .particle_filter import *